Add feature-specific themes.
The app's main (global) theme is left untouched. This commit adds the themes to
the appropriate pages. Ultimately, these themes will also be applied to the
corresponding cards on the overview table as well.

This adds themes for the 'accounts', 'bills', and 'budgets' cards, but does NOT
add anything for the greyscale 'alerts' card (which does not seem to anything
beyond what is provided by the root mat-dark-theme).